A ball on a parachute
* Add Draw_Parachute_In_Level method
* Add two-color color to AColor
* Add Draw_Parachute_Part() method
* Add Set_On_Parachute() method
* Add EBS_On_Parachute state
* Add Draw_Parachute() method
* Add Parachute_Rect
* Realize the movement of the ball on the parachute
* Realize parachute disconnection when hitting the platform
* Realize EBS_Off_Parachute handling in ABall::Set_State()
* Add method Redraw_Parachute()
* Add Clear_Parachute() method
* Realize EBS_On_Parachute state setting prohibition in ABall:Set_State() method
* Realize ball loss on parachute

+++ Popcorn/Active_Brick.cpp
@@ -249,7 +249,7 @@ void AActive_Brick_Multihit::Draw(HDC hdc, RECT &paint_area)
 	xform.eM12 = 0.0f;
 	xform.eM21 = 0.0f;
 	xform.eM22 = 1.0f;
-	xform.eDx = (float)Brick_Rect.left + (1.0 - x_ratio) * (float)(AsConfig::Brick_Width * AsConfig::Global_Scale) / 2.0f;
+	xform.eDx = (float)Brick_Rect.left + (float)(1.0 - x_ratio) * (float)(AsConfig::Brick_Width * AsConfig::Global_Scale) / 2.0f;
 	xform.eDy = (float)Brick_Rect.top;
 	GetWorldTransform(hdc, &old_xform);
 	SetWorldTransform(hdc, &xform);
